added separate component for DeleteItemModal, fixed translations
