+++ src/modules/Builder/features/ActivityItems/DeleteItemModal/DeleteItemModal.types.ts
@@ -1,0 +1,8 @@
+import { Dispatch, SetStateAction } from 'react';
+
+export type DeleteItemModalProps = {
+  itemIdToDelete: string;
+  setItemIdToDelete: Dispatch<SetStateAction<string>>;
+  activeItemIndex: number;
+  setActiveItemIndex: Dispatch<SetStateAction<number>>;
+};
