+++ src/modules/Builder/features/ActivityItems/DeleteItemModal/index.ts
@@ -1,0 +1,1 @@
+export * from './DeleteItemModal';
