+++ src/modules/Builder/features/ActivityItems/ItemConfiguration/OptionalItemsAndSettings/SkippedItemInVariablesModal/SkippedItemInVariablesModal.types.ts
@@ -1,5 +1,3 @@
 export type SkippedItemInVariablesModalProps = {
   itemName: string;
 };
-
-export type ItemNameWIthIndex = { index: number; name: string };
