+++ src/resources/app-en.json
@@ -875,8 +875,8 @@
     "incorrectEmail": "Email must be valid",
     "variableInTheSameItem": "* You cannot use item name in the same item. Please remove",
     "variableIsNotSupported": "* This item is not supported, please remove it.",
-    "variableReferringToSkippedItem": "Remove the variable referring to the nonexistent item.",
-    "variableReferringToNotExistedItem": "Remove the variable referring to the not existed item."
+    "variableReferringToSkippedItem": "Remove the variable referring to the skipped item.",
+    "variableReferringToNotExistedItem": "Remove the variable referring to the nonexistent item."
   },
   "apply": "Apply",
   "addNewNote": "Add a New Note",
